+++ NEWS.md
@@ -10,6 +10,7 @@ editor_options:
 -   Revise `stat_multcomp()` to flexibly include only when using "letters" an additional label with the critical _P_-value and the method used to adjust the empirical _P_-values (suggested by _markbneal_).
 -   Revise `stat_multcomp()` to support arbitrary sets of pairwise contrasts.
 -   Rename parameters `adjusted.type` to `p.adjust.method`, and `constrast.type` to `contrasts` of `stat_multcomp()`.
+-   Change defaults used for `p.adjust.method` so that they depend on the argument passed to `contrasts`.
 
 # ggpmisc 0.5.5